requirements.txt
easysnmp==0.2.5
python-dotenv~=0.16.0
tabulate~=0.8.9